Refactor to Feedback page to FeedbackForm component
